+++ examples/inner_html/README.md
@@ -1,0 +1,8 @@
+# Inner HTML Example
+
+This example renders unescaped HTML by manually handling the DOM element.
+
+## Concepts
+
+- Manually creating `Html` without the `html!` macro.
+- Using `web-sys` to manipulate the DOM.
