+++ examples/inner_html/index.html
@@ -1,0 +1,9 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="utf-8" />
+    <title>Yew • Inner HTML</title>
+  </head>
+
+  <body></body>
+</html>
