--- examples/inner_html/src/lib.rs
@@ -1,49 +1,0 @@
-#![recursion_limit = "512"]
-
-use web_sys::{console, Node};
-use yew::virtual_dom::VNode;
-use yew::{Component, ComponentLink, Html, ShouldRender};
-
-const HTML: &str = include_str!("document.html");
-
-pub struct Model {
-    pub value: i64,
-}
-
-pub enum Msg {}
-
-impl Component for Model {
-    type Message = Msg;
-    type Properties = ();
-
-    fn create(_: Self::Properties, _: ComponentLink<Self>) -> Self {
-        Model { value: 0 }
-    }
-
-    fn update(&mut self, _: Self::Message) -> ShouldRender {
-        true
-    }
-
-    fn change(&mut self, _: Self::Properties) -> ShouldRender {
-        false
-    }
-
-    fn view(&self) -> Html {
-        let js_svg = {
-            let div = web_sys::window()
-                .unwrap()
-                .document()
-                .unwrap()
-                .create_element("div")
-                .unwrap();
-            div.set_inner_html(HTML);
-            console::log_1(&div);
-            div
-        };
-        eprintln!("js_svg: {:?}", js_svg);
-        let node = Node::from(js_svg);
-        let vnode = VNode::VRef(node);
-        eprintln!("svg: {:?}", vnode);
-        vnode
-    }
-}
